+++ src/theme/components/progressIndicator.ts
@@ -49,6 +49,7 @@ const getCircularContainerFlexDir = (labelPlacement) => {
 const ProgressIndicator = defineMultiStyleConfig({
   baseStyle: definePartsStyle(
     ({ darkMode, size, labelPlacement }: ProgressIndicatorBaseStyle) => {
+      const circularLabelMargin = size === "default" ? "xs" : "xxs";
       return {
         color: darkMode
           ? "dark.ui.typography.heading"
@@ -85,10 +86,10 @@ const ProgressIndicator = defineMultiStyleConfig({
           width: "fit-content",
         },
         circularLabel: {
-          marginBottom: labelPlacement === "top" ? "xxs" : 0,
-          marginLeft: labelPlacement === "right" ? "xxs" : 0,
-          marginRight: labelPlacement === "left" ? "xxs" : 0,
-          marginTop: labelPlacement === "bottom" ? "xxs" : 0,
+          marginBottom: labelPlacement === "top" ? circularLabelMargin : 0,
+          marginLeft: labelPlacement === "right" ? circularLabelMargin : 0,
+          marginRight: labelPlacement === "left" ? circularLabelMargin : 0,
+          marginTop: labelPlacement === "bottom" ? circularLabelMargin : 0,
           fontSize:
             size === "default"
               ? "desktop.label.label1"